Type fixing for numpy update.

+++ pyearth/_types.pyx
@@ -1,5 +1,6 @@
+# cython: language_level=3
 import numpy as np
-FLOAT = np.float64
-INT = np.int
-INDEX = np.intp
+FLOAT = float
+INT = int
+INDEX = int
 BOOL = np.uint8
